add modules feature_settings, feature_shoplist. delete unused from app, refactoring
