+++ app/src/main/java/com/example/shoplist/ui/MainActivity.kt
@@ -15,12 +15,19 @@ import com.example.shoplist.feature_favorites.ui.FavoritesFragment
 import com.example.shoplist.feature_meal_item.ui.MealsViewHolder
 import com.example.shoplist.feature_recipes.ui.RecipeFilterFragment
 import com.example.shoplist.feature_recipes.ui.RecipesFragment
+import com.example.shoplist.feature_settings.models.Themes
+import com.example.shoplist.feature_settings.ui.SettingsFragment
+import com.example.shoplist.feature_shoplist.ui.ShoplistFragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 private const val SHARED_PREFERENCES_NAME = "settings"
 
-class MainActivity : AppCompatActivity(), SettingsFragment.Controller, MealsViewHolder.Contract, RecipeFilterFragment.Contract {
-    private val binding by viewBinding(ActivityMainBinding::bind , R.id.activity_container)
+class MainActivity : AppCompatActivity(),
+    SettingsFragment.Controller,
+    MealsViewHolder.Contract,
+    RecipeFilterFragment.Contract {
+
+    private val binding by viewBinding(ActivityMainBinding::bind, R.id.activity_container)
     private val bottomNavigationView: BottomNavigationView by lazy { binding.bottomNavigationView }
     private lateinit var sharedPreferences: SharedPreferences
     private val sharedValueNameTheme = "Theme"
@@ -35,8 +42,10 @@ class MainActivity : AppCompatActivity(), SettingsFragment.Controller, MealsView
         }
         setContentView(R.layout.activity_main)
 
-        bottomNavigationView.setOnItemSelectedListener { item -> setNavigation(item) }
-        bottomNavigationView.selectedItemId = R.id.nav_recipes
+        bottomNavigationView.apply {
+            setOnItemSelectedListener { item -> setNavigation(item) }
+            selectedItemId = R.id.nav_recipes
+        }
     }
 
     private fun setNavigation(item: MenuItem): Boolean {
@@ -49,16 +58,16 @@ class MainActivity : AppCompatActivity(), SettingsFragment.Controller, MealsView
     }
 
     private fun openFragment(fragmentInstance: Fragment, backstack: Boolean) {
-        if (backstack) {
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.container, fragmentInstance)
-                .addToBackStack(null)
-                .commit()
-        } else {
-            supportFragmentManager.beginTransaction()
-                .replace(R.id.container, fragmentInstance)
-                .commitNow()
+        supportFragmentManager.beginTransaction().apply {
+            replace(R.id.container, fragmentInstance)
+            if (backstack) {
+                addToBackStack(null)
+                commit()
+            } else {
+                commitNow()
+            }
         }
+
     }
 
     override fun onCreateOptionsMenu(menu: Menu?): Boolean {
