--- app/src/main/java/com/example/shoplist/ui/ShoplistFragment.kt
@@ -1,11 +1,0 @@
-package com.example.shoplist.ui
-
-import androidx.fragment.app.Fragment
-import com.example.shoplist.R
-
-class ShoplistFragment : Fragment(R.layout.fragment_shoplist) {
-
-    companion object {
-        fun newInstance() = ShoplistFragment()
-    }
-}
