--- app/src/main/java/com/example/shoplist/ui/Utils.kt
@@ -1,5 +1,0 @@
-package com.example.shoplist.ui
-
-enum class Themes {
-    DEFAULT, INDIGO, ORANGE
-}
