--- app/src/main/res/layout/fragment_settings.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_margin="@dimen/default_margin"
-    tools:context=".ui.SettingsFragment">
-
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/theme_title_text_view"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="@dimen/default_padding"
-        android:text="@string/settings_theme_title_text"
-        android:textAlignment="center"
-        android:textSize="@dimen/large_text_size"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <RadioGroup
-        android:id="@+id/settings_themes_radio_group"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintTop_toBottomOf="@+id/theme_title_text_view" >
-
-        <com.google.android.material.radiobutton.MaterialRadioButton
-            android:id="@+id/settings_themes_radio_default"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/theme_default"/>
-
-        <com.google.android.material.radiobutton.MaterialRadioButton
-            android:id="@+id/settings_themes_radio_indigo"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/theme_indigo" />
-
-        <com.google.android.material.radiobutton.MaterialRadioButton
-            android:id="@+id/settings_themes_radio_red"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="@string/theme_orange"/>
-    </RadioGroup>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
