+++ feature_settings/src/main/java/com/example/shoplist/feature_settings/models/Themes.kt
@@ -1,0 +1,5 @@
+package com.example.shoplist.feature_settings.models
+
+enum class Themes {
+    DEFAULT, INDIGO, ORANGE
+}
