+++ feature_shoplist/src/main/java/com/example/shoplist/feature_shoplist/ui/ShoplistFragment.kt
@@ -1,0 +1,11 @@
+package com.example.shoplist.feature_shoplist.ui
+
+import androidx.fragment.app.Fragment
+import com.example.shoplist.feature_shoplist.R
+
+class ShoplistFragment : Fragment(R.layout.fragment_shoplist) {
+
+    companion object {
+        fun newInstance() = ShoplistFragment()
+    }
+}
