+++ settings.gradle
@@ -19,3 +19,5 @@ include ':feature_recipes'
 include ':feature_meal_item'
 include ':feature_favorites'
 include ':data_local'
+include ':feature_settings'
+include ':feature_shoplist'
